Create License
Copyright (c) 2025 Pranav Krishnakant Bhavsar  
All Rights Reserved.  

This software and its associated files are the exclusive property of Pranav Krishnakant Bhavsar.  
Unauthorized copying, modification, distribution, or use of this software, in whole or in part,  
is strictly prohibited without prior written permission from the copyright owner.  

This software is provided "as is" without any warranties, express or implied, including  
but not limited to warranties of merchantability or fitness for a particular purpose.  

For inquiries regarding licensing, please contact: aries200326@gmail.com